requirements.txt
# COEVOLVE Framework Dependencies

# Core LLM Providers
openai>=1.0.0
anthropic>=0.7.0

# Agent Orchestration & Multi-Agent Frameworks
langchain>=0.1.0
langgraph>=0.0.20
autogen-agentchat>=0.2.0  # Microsoft AutoGen for multi-agent
crewai>=0.1.0  # Alternative multi-agent framework

# Memory & Vector Stores
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pinecone-client>=2.2.0
weaviate-client>=3.25.0

# ML & Training Frameworks
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.25.0
peft>=0.6.0  # For LoRA/DPO training
trl>=0.7.0  # Transformer Reinforcement Learning (for DPO/PPO)
bitsandbytes>=0.41.0  # For quantization

# Local Model Serving & Inference
vllm>=0.2.0  # Fast inference server
llama-cpp-python>=0.2.0  # Local Llama models
text-generation>=0.6.0  # HuggingFace text generation

# Experiment Tracking & Monitoring
wandb>=0.16.0  # Weights & Biases for experiment tracking
mlflow>=2.8.0  # MLflow for model registry and tracking
tensorboard>=2.14.0  # TensorBoard for visualization

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Data Handling & Validation
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
jsonschema>=4.19.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
rich>=13.6.0  # Beautiful terminal output
typer>=0.9.0  # CLI framework

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
pandas>=2.0.0

# Code Analysis & Security (for Red Team/Blue Team)
bandit>=1.7.5  # Security linter
pylint>=3.0.0  # Code quality
radon>=6.0.1  # Code complexity metrics

# Notebooks & Interactive Development
jupyter>=1.0.0
ipywidgets>=8.1.0
jupyterlab>=4.0.0

# Testing & Quality Assurance
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0

# Async & Concurrency
asyncio>=3.4.3
aiohttp>=3.9.0
uvloop>=0.19.0  # Faster event loop

# Graph & State Machine Libraries
networkx>=3.1  # For debate trees and skill graphs
graphviz>=0.20.1  # Visualization of state machines

# Database & Persistence
sqlalchemy>=2.0.0  # For structured data storage
redis>=5.0.0  # For caching and message queues

# API Frameworks (for serving models)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0

# Prompt Engineering & LLM Tools
guidance>=0.1.0  # Microsoft Guidance for structured generation
outlines>=0.0.20  # Structured text generation
langfuse>=2.0.0  # LLM observability

# Evaluation & Benchmarking
bert-score>=0.3.13  # For text similarity
rouge-score>=0.1.2  # For summarization eval
sacrebleu>=2.3.1  # For translation eval

# Optional Advanced Features
# openai-whisper>=20230918  # For speech capabilities
# elevenlabs>=0.2.0  # For text-to-speech
# stable-baselines3>=2.1.0  # If using traditional RL alongside LLM RL
